requirements3.txt
-e git+https://github.com/PostDispatchInteractive/authomatic.git@7ff4cba0ac0773cc547e7ed9a4ab06ccda346c5c#egg=Authomatic
awscli==1.18.137
bcdoc==0.16.0
boto==2.49.0
botocore==1.17.60
colorama==0.4.3
copytext==0.2.1
cssmin==0.2.0
docutils==0.15.2
ecdsa==0.16.0
Fabric3==1.14.post1
facebook-sdk==3.1.0
Flask==1.1.2
google-api-python-client==1.11.0
gunicorn==20.0.4
htmlmin==0.1.12
httplib2==0.18.1
jdcal==1.4.1
Jinja2==2.11.2
jmespath==0.10.0
MarkupSafe==1.1.1
nose==1.3.7
odict==1.7.0
openpyxl==3.0.5
paramiko==2.7.2
ply==3.11
pyasn1==0.4.8
pycrypto==2.6.1
python-dateutil==2.8.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.5.0
six==1.15.0
smartypants==2.0.1
ssh==1.8.0
termcolor==1.1.0
twitter==1.18.0
Werkzeug==1.0.1
